+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.1] - 2023-04-17
 
 - Fix for `other_keywords` to accept the optional type as intended.
+- Fix for inheritance issues when including VCDry::DSL on a parent component.
 
 ## [1.0.0] - 2023-04-16